+++ zio-http-cli/src/main/scala/zio/http/endpoint/cli/description.scala
@@ -1,0 +1,5 @@
+package zio.http.endpoint.cli
+
+import scala.annotation.StaticAnnotation
+
+class description(val text: String) extends StaticAnnotation // TODO move this to zio-schema
